update sore tgl 17

+++ database/migrations/2022_12_15_043055_create_pembayaran_table.php
@@ -15,7 +15,7 @@ class CreatePembayaranTable extends Migration
     {
         Schema::create('pembayaran', function (Blueprint $table) {
             $table->engine = 'InnoDB';
-            $table->unsignedInteger('id_pembayaran');
+            $table->unsignedBigInteger('id_pembayaran');
             $table->primary('id_pembayaran');
             $table->date('tanggal_bayar');
             $table->double('total_bayar');
